# julio-cesar-ipv6-converter-api-1.0.0-swagger.yaml
openapi: 3.0.0
info:
  version: '1.0.0'
  title: 'IPv6 Converter with algorithm EUI-64 - API Specification'
  description: 'IPv6 Converter with algorithm EUI-64 - API Specification'

security:
  - application:
    - read
    - write

paths:
  /api/v1/convert_ipv6:
    parameters:
      - in: query
        name: ipv6_prefix
        schema:
          type: string
          example: '2000:ffff:ffff:ffff'
        description: The IPv6 prefix
      - in: query
        name: mac
        schema:
          type: string
          example: 'ff-ff-ff-ff-ff-ff'
        description: The MAC Address

    get:
      summary: Get the whole IPv6 formed with the given IPv6 prefix and Mac Address
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/body_response_convert'
        400:
          description: Bad Request
        500:
          description: Internal Server Error

  /api/v1/valid_ipv6:
    parameters:
      - in: query
        name: ipv6
        schema:
          type: string
          example: '2000:ffff:ffff:ffff:fdff:ffff:feff:ffff'
        description: The IPv6

    get:
      summary: Get the validation for the given IPv6
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/body_response_valid'
        400:
          description: Bad Request
        500:
          description: Internal Server Error

  /api/v1/valid_ipv4:
    parameters:
      - in: query
        name: ipv4
        schema:
          type: string
          example: '192.255.255.255'
        description: The IPv4

    get:
      summary: Get the validation for the given IPv4
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/body_response_valid'
        400:
          description: Bad Request
        500:
          description: Internal Server Error

  /api/v1/valid_mac:
    parameters:
      - in: query
        name: mac
        schema:
          type: string
          example: 'ff-ff-ff-ff-ff-ff'
        description: The MAC Address

    get:
      summary: Get the validation for the given MAC Address
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/body_response_valid'
        400:
          description: Bad Request
        500:
          description: Internal Server Error

components:
  schemas:
    body_response_convert:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: The IPv6 formed with the given IPv6 prefix and Mac
          example: '2000:ffff:ffff:ffff:fdff:ffff:feff:ffff'
    body_response_valid:
      type: object
      required:
        - message
      properties:
        message:
          type: boolean
          description: The boolean that indicates if the given parameter is valid or not
          example: True